src/api/recommendation_algorithm.py
class RecommendationAlgorithm:
    """
    A class to implement the recommendation algorithm for suggesting players based on user preferences.
    """

    def __init__(self, user_preferences, player_data):
        """
        Initializes the RecommendationAlgorithm with user preferences and available player data.
        """
        self.user_preferences = user_preferences
        self.player_data = player_data

    def compute_recommendations(self):
        """
        Computes recommendations using cosine similarity based on user preferences.
        """
        recommendations = []
        # Calculate cosine similarity here
        # Add logic to generate recommendations based on self.user_preferences
        return recommendations

    def fetch_user_preferences(self, user_id):
        """
        Fetches user preferences based on user ID.
        """
        # Logic to retrieve user preferences from a database
        return self.user_preferences